Change downloading of episodes now uses AFNetworking
To start downloading episodes use downloadEpisodeWithURL:saveToURL:downloadProgress:success:failure: method located in IGHTTPClient. The download method also make use of AFNetwork's built in network reachability for checking network status. This also address issue #11 a local notification is now fires if the app is in background state and a download succeeds or fails.