Refactor the download methods to remove dependence on the table view cell and support resuming of download if episode already is partial downloaded
Added AFDownloadRequestOperation which is now used to download the episodes instead of AFHTTPRequestOperation because it has built in support for resuming partially downloaded data.

Removed the downloadProgress block from IGHTTPClient downloadEpisodeWithURL:targetPath:success:failure because it's not needed.

IGEpisode isCompletelyDownloaded now checks to see if the episode exists in the Episodes directory and isPartiallyDownloaded checks to see if a temp file exists in the Incomplete directory.

IGEpisodeTableViewCell now sets the downloadStatus to IGEpisodeDownloadStatusNotDownloading if download operation is paused.