Change now playing info set manually not using the common meta data from the episode file
On some episodes the ID3 tag data which was used to set the now playing info was different or just not there. A more better idea would be to edit the ID3 tag data after the episode file has been downloaded but this would take too long. Might include in another version. For now this change should work fine.