Add Issue #3 Support for video episodes
The video episodes are hosted on Vimeo and YouTube so I have put them into a separate feed. The IGVideoPlayerViewController is just a UIWebView that loads the web page where the video is hosted. Video playback in landscape is supported, so I added landscape navigation bar background and button images. Downloading video episodes is not supported.